separated ListboxOption once again, added scroll meta test
